refactor(cli): organize command files into a dedicated package
